Move dbconnect to util package
